requirements.txt
torch
torchvision
pytorch-lightning
numpy
pandas
scikit-learn
matplotlib
seaborn
wandb
tqdm
jupyter
pyyaml
torchmetrics
optuna
